Allow giving a secondary order which shouldn't interfere with the primary.
Okay, that is a crap description.

If you want to say SELECT foo FROM bar ORDER BY location DESC, name ASC; the
AddOrder for name would have become the default order and would have appeared
in any URLs for the page.  Whereas location is really the want you'd want
to propogate in the URLs.

By giving the secondary parameter to AddOrder it won't globber the state
variables and therefore won't propogate.