[XMLDocument] Fix for NS helpers assigning used NS alias.
The namespace alias assigned by helper methods like CalDAVElement()
and CalendarServerElement() attempt to assign a preferred alias
but if this has already been used for a different namespace the
collision would error.

This fix simply gives up on the preferred alias and lets one be
auto-assigned within the AddNameSpace() method.