[XMLDocument] Return any created node. Also parse namespace in tag.
These functions are much more useful if we can actually use any
returned reference as another location within a document, so we
can have a hold of the root, and then add within each new tag
we add, supporting outer -> inner construction approaches along
with the existing inner -> outer approach used currently.

Additionally the NewXMLElement method is changed to parse the
namespace out of the tag, as the NSElement already does.